requirements.txt
anthropic
fastapi
uvicorn
python-dotenv
tavily-python